Growth_Mindset_Challenge/requirements.txt
streamlit
pandas
seaborn
scikit-learn
python-dotenv
google-generativeai
openpyxl  # Required for reading/writing Excel files
pyarrow  
